+++ src/main/java/ua/foxminded/pinchuk/javaspring/carrestservice/service/BrandService.java
@@ -5,7 +5,7 @@ import ua.foxminded.pinchuk.javaspring.carrestservice.entity.Brand;
 import java.util.List;
 
 public interface BrandService {
-    Brand findById(int id) throws Exception;
+    Brand findById(Long id) throws Exception;
     void saveOrUpdate(Brand brand);
     void remove(Brand brand);
     List<Brand> findAll();
