+++ src/main/java/ua/foxminded/pinchuk/javaspring/carrestservice/service/TypeService.java
@@ -5,7 +5,7 @@ import ua.foxminded.pinchuk.javaspring.carrestservice.entity.Type;
 import java.util.List;
 
 public interface TypeService {
-    Type findById(int id) throws Exception;
+    Type findById(Long id) throws Exception;
     void saveOrUpdate(Type type);
     void remove(Type type);
     List<Type> findAll();
